Piano Communication Thread
Before, the piano midi messages processing was within the checking and counting segment of the code, which caused many of the midi information to be missed during other process. Now the piano communication processing is done in its own thread, making midi data loss very unlikely. Still need to implement a method to assure that losing midi data is extremely unlikely.